Projection works with new method
The projection onto u^0 is not needed anymore, now it works nicely.
When I project onto C-space the Tr(Au^i) = 0!

The init_Z works too, the result of Tr(Zu^i) = h^i, the hamiltonian.
Now I have to look for a way to make sure it's always positive.